Revert "Remove equality-only restriction on input reader filters"
Summary:
From Brett Slatkin: "The restriction is Megastore. You can only ever have one inequality filter in megastore, but as many equality filters as you want. The inequality filter used for mapreduce is on Key, to split the entities into approximately equal-sized ranges. In theory you could move this to do splits on a datetime field, but it may be problematic because any equal date fields would cause the split points to get fuzzy, where as all keys are guaranteed to be unique, so offsetting in a list of keys is always reliable."

This reverts commit 9adca192251bc70f97859b5d16e1e7316b3f1f61.

Ignore *.py[co]

Do not lint anything; this is a third-party mirror

Test Plan: None needed, this reverts the third-party library to a known-good state

Reviewers: csilvers

Reviewed By: csilvers

CC: chris, kamens

Differential Revision: http://phabricator.khanacademy.org/D646