Add in log messages when quota is low or gone
Summary:
Since the quota manager relies on memcache, some recent mapreduce
failures may be a result of the quota values being flushed from the
cache. It appears the code is currently structured so that the entity
scans are aborted when the quota is low or not in memcache. That seems
undesirable, but first we need to verify that that's the actual issue.

Test Plan: Ran the send_welcome_emails locally using dev_appserver.py and --backends, and verified that nothing explodes. I did not explicitly attempt to mock out quota/memcache values so they were low, though.

Reviewers: csilvers, chris

Differential Revision: http://phabricator.khanacademy.org/D757