Use filters, when possible, for deciding on keys to split.
Summary:
The default algoritym for picking keys to split basically -- I'm
simplifying -- picks num-shards random keys from all the keys in the
entity we're running the mapreduce over.  (Using the __scatter__
property, which assigns random values to a subset of all keys.)

When filters were introduced, this made the key-splitting logic
diverge from the input-selection logic, since the input-selection uses
the filters but the key-splitting doesn't.  If the filters are
non-uniform, it means bad key-splitting. (For instance, if you have a
User model where the keys are people's names, and your filter is
"last_name starts with s", then the key-splitting logic might split the
keys as (a..e, f..j, k..o, p..t, u..z) when it should do (a..se,
sf..sj, sk..so, sp..st, su..z).)

This change adds filters into the key-splitting logic as well.  This
resolves the divergence, and should speed up mapreduces -- in our
tests by a factor of 2 -- since keys will be distributed to workers
more evenly.

The cost is that you need to introduce a new db-index, which is 'your
filter + __scatter__.'  This code checks if such an index exists, and
if so uses it, and if not falls back on the old, filter-less splitting
behavior.  (It logs the db-index to add, for folks who would like.)

Test Plan:
Ran the new code manually to make sure it gave reasonable results on
the KA UserData model with last_modified_as_mapreduce_epoch == 4:
   % tools/devshell.py --prod
   >>> from google.appengine.api import datastore
   >>> import user_models
   >>> ds_query = datastore.Query(kind='UserData', keys_only=True)
   >>> ds_query.Order("__scatter__")
   >>> random_keys = ds_query.Get(72)
   >>> random_keys.sort()
   >>> random_keys[:10]
and noted that without the filter, the first 7 keys were all of the
form 'user_email_key_...', which are all obsolete users (and thus have
last_modified_as_mapreduce_epoch == 0).  Then I did:
   >>> ds_query = datastore.Query(kind='UserData', keys_only=True)
   >>> ds_query.update({'last_modified_as_mapreduce_epoch =': 4})
   >>> ds_query.Order("__scatter__")
   >>> random_keys = ds_query.Get(72)
   >>> random_keys.sort()
   >>> random_keys[:10]
and verified none of the random keys had the form
'user_email_key_...'.  They were all compressed into the first bucket,
where they'll be ignored, just like we want.

An actual 'successful faster mapreduce' test will have to wait until
we push this change to production and try a new mapreduce.

Reviewers: chris

Reviewed By: chris

CC: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D781