Support reading the data files even when mapreduce is in a zipfile.
Summary:
zipimport makes it possible to read python files from a .zip, but it
doesn't address reading data files.  The code has to be written to
support that.  Change the mapreduce code so it is.  Luckily it's
pretty localized.

Don't assume that zipfiles end with .zip, and .zip-things are zipfiles.

Just let the zipfile module tell us whether something is a zipfile or not.

Test Plan:
Did a non-standard deploy and visited the following urls and made sure
they gave back non-error content:
   http://znd-zipimport.khan-academy.appspot.com/mapreduce/status
   http://znd-zipimport.khan-academy.appspot.com/mapreduce/status.js
   http://znd-zipimport.khan-academy.appspot.com/mapreduce/jquery.js
   http://znd-zipimport.khan-academy.appspot.com/mapreduce/lib/pipeline/status.js
   http://znd-zipimport.khan-academy.appspot.com/mapreduce/lib/pipeline/images/treeview-black.gif

Reviewers: alpert, kamens

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D884