Add logging.info() statements to all of the pipeline handlers
Summary:
I have been seeing this very weird hanging problem when pipeline jobs are launched.  Here is the experiment and observations I made tonight:

1. Set max_concurrent_requests: 10 on pipeline-reqeusts-queue, also set version = znd-pipeline-tests-highmembackend
2. Execute the 47 Idaho coaches job from https://www.khanacademy.org/devadmin/studentusedata
3. Let it sit for ~45 minutes.  Eventually, it enters this hung state which can be described as:

- The pipeline-report-queue lists "10 running tasks" and ~1500 queued tasks https://appengine.google.com/queuedetails?queue_name=pipeline-report-queue&app_id=s~khan-academy
- The -highmembackend instances page shows 10 highmembackend servers that are currently running, and have been running for the entire 45 minutes
- The logs UI on the dashboard shows no new log entires being written

This makes me think that the task queue has fired a request to 10 of the highmembackends, and they have simply not responded, and not logged any information as to what they are doing.  I looked through the pipeline code a bit more and noticed a few vague opportunities for exception swallowing or infinite looping, but without knowing what the highmembackend machines are doing in this hung state, I have no way to investigate further.  If only I could SSH into the machine and get a memory dump. Then I could get a stack trace of all the currently running threads.  That's what I would have done on the outlook.com servers, but noooo. </rant>

Anyway, I added this logging and re-ran the job, and the hang did not reproduce. So, I would like to check this in permenantly so next time this happens we can get a little better idea of which handlers are causing the backends to hang.

Test Plan:
Deploy as a znd- run a job from https://www.khanacademy.org/devadmin/studentusedata
Watch the log output

Reviewers: jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D4963