Don't fail in the admin UI when child pipelines are missing
Summary:
In the pipeline admin UI, child pipelines can be missing because the query used
is eventually consistent, and the UI seems to work just fine even with some
pipelines missing, so we should just display results on a best-effort basis.

Test Plan: Ran the summarize task with these changes. Things seem to work.

Reviewers: mattfaus

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D7702