Truncate slot values in addition to pipeline args in the admin UI
Summary:
BigQuery returns the SQL text in its job status, which was a pipeline return
value, so I was getting admin UI results on the order of 30MB, even though some
values are truncated. This change truncates return values as well, so the
response size is much more reasonable.

Test Plan:
Run the summarizer task and load the admin UI. Things seem to work, and slot
return values are truncated, as expected.

Reviewers: mattfaus

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D7748