Add a "Load full values" option to the admin UI
Summary:
Parameter and output values are truncated to avoid hitting response size limits,
but they're sometimes useful when debugging individual pipelines. This commit
adds a "Load full values" link to every pipeline page that does an ajax request
to get the full argument and output slot values, then replaces the truncated
values with the full values in the status map and re-renders the page.

Test Plan:
-Open up a root pipeline that ran a bunch of large SQL statements. Make sure
it's possible to load the full values in both parameters and outputs.
-Go to a partially-loaded pipeline page where some output slots haven't been
filled yet and try loading those. It correctly loads the arguments.
-Stop dev_appserver with the page still open and try clicking "Load full
values". It fails and gives a retry link.
-Start up dev_appserver again and click the retry link. It succeeds.

Reviewers: mattfaus

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D7845