Add a None-check when retrieving _BarrierListener entities
Today, I had a pipeline job (https://khanacademy.org/_ah/pipeline/status?root=865d4307bb6711e3b57113246976b5f0) that was hung because this code was throwing an exception. The symptom was that there was a task in my queue with many retries (18), so I found the exception that task was raising:

```
  File "/base/data/home/apps/s~khan-academy/highmem:znd-sud.374884420765968787/third_party/mapreduce/lib/pipeline/pipeline.py", line 2493, in post
    self.request.get('cursor'))
  File "/base/data/home/apps/s~khan-academy/highmem:znd-sud.374884420765968787/third_party/mapreduce/lib/pipeline/pipeline.py", line 1475, in notify_barriers
    blocking_slot_keys.extend(barrier.blocking_slots)
AttributeError: 'NoneType' object has no attribute 'blocking_slots'
```

This makes sense because the code that creates these records is a non-atomic write.

Auditors: alan

Test Plan:
After making this change to the code, I overwrote the deployed code (because I was doing a znd-deploy) and clicked "Run now" on the task that was infinitely retrying. The task continued successfully, and my pipeline job ran to completion.