Fix occasional crash due to missing _LowMemoryPipelineRecord.key
Summary:
_LowMemoryPipelineRecord needs to behave like a _PipelineRecord for all of our
use cases, and I originally thought that the key was never used, but it turns
out that it's used in a warning message in get_status_tree. This commit adds the
method in the straightforward way, which fixes an occasional crash in the
pipeline UI.

Test Plan:
Copy the warning message to an earlier place so it runs every time. Run the
code without this fix, and it consistently crashes. Apply this fix and it
consistently loads the UI correctly.

Reviewers: mattfaus, chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D7965