Don't use the current version as the default target for root pipelines
Summary:
This is a necessary step to avoid the next commit causing problems. The code
seems to have an attempt to make it so the module and version of spawned
pipelines are the same as the code that spawned them, but there's a bug where
this "target" value is being ignored most of the time. If I just fix that bug,
though, then all top-level pipeines spawned from the default module will stay on
the default module, which isn't what we want. So, this change makes it so when a
top-level pipeline is created, we leave the target as None (meaning we use
whatever target is used by the queue) unless a target was assigned explicitly.

Test Plan:
Run /admin/bigbingo/summarize_end_to_end_test . The root pipeline has an
"Unspecified" target and every other pipeline has the regular target.
Temporarily change the end to end test target to "foo" and run
/admin/bigbingo/summarize_end_to_end_test . The top-level target is "foo" and
the rest are the regular target.

Reviewers: chris, mattfaus

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D8587