Revert "Don't use the current version as the default target for root pipelines"
Summary:
This reverts commit c62b5a2b1cb28a137d097d06d70abed6c08ff4bc. I didn't realize
when writing that commit that task queue targets would override explicit
targets. This means that if the task queue had a target, then my change had no
effect anyway, and if the task queue didn't have a target, then the "None" value
that I change the target to makes it so the task runs on the current version
anyway, so there's no point in setting it back to None.

In addition to the revert, I also documented this issue a little better.

Test Plan:
Run /admin/bigbingo/summarize_end_to_end_test and sanity-check the values. Then
remove the target specified there and sanity-check the values.

Reviewers: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D8665