Ignore disabled LogService API when looking for previous shard slice state
When running a mapreduce handler on a managed VM, the LogService API is not yet supported, so trying to fetch the previous shard slice's state throws an exception and fails the mapreduce job. Instead, we should handle this situation more gracefully and just assume that the old request did not finish.

Test Plan:
Deploy this code to my Managed VM test module and run the test mapreduce job.

Auditors: jace